Separete bugle into multiple files
